+++ CHANGELOG.md
@@ -2,3 +2,4 @@ v1.0.1 -> v1.0.2
 
 - fixed military time setting
 - added background presets
+- added animation speed setting